Add transfer methods to TraderClient and global mock-Trader
Trader now implements an observer-based transfer-notification mechanism.
TransferReceivedEvents will get put in all the according listeners’ queues.

A TransferReceivedListener transforms those Events into a TransferReceipt, which
additionally contains the timestamp when the Transfer was received for internal handling.

According methods are mirrored in the client/server parts of the trader but aren’t tested yet.