Add Commitment-Service-node logic, object and tasks
The CommitmentService acts on SwapCommitments, which are Object-oriented containers for all the necessary information of the state of an AssetSwap from the CommitmentServices perspective.

The SwapCommitment-object accepts specific Messages and TransferReceipts and stores them in private members.
It will hand out messages and Refund-Object, based on the state of the Swap.
If a query for such a handout is not valid, one will receive a `False`-return value.

The business-logic is processed by greenlet-tasks, that listen for specific messages on the message_broker and on incoming transfers at the trader-client.

If such an event is received, the task will retrieve the according SwapCommitment with the offer-id / transfer-identifier associated with that swap and perform actions on it.

The MessageSenderTask and RefundTask will handle the outgoing messages and transfers with the message_broker-client and trader-client accordingly.
Other tasks will put messages and Refund-Objects in queues, and the Sender- and Refund tasks will blindly consume and forward them.