Implement the CommitmentService-interface for commits in the Raidex-node
This will handle all the communication and transfer-logic with the CommitmentService from the perspective of the Raidex-node.

It offers a minimal interface for conducting maker- and taker-commits,
so that the caller only needs to provide the associated Offer-object.
The CommitmentService handles the logic and returns the ProvenOffer on
success (asynchronously)