Add integration test for the Node to CS interaction
This test involves a scenario where 3 actors on the market are trying to handle the commitments with the commitment-service in order to trade securely.
The interaction is only indirect through the message-broker and the commitment-transfers of the traders. The actual AssetSwap will be omitted and the actors just notify the network that the Swap was executed.