Improve RaidexNode’s default constructor, Mark MockExchangeTask outdated
The default class method constructor of the RaidexNode now only takes 
arguments as they would be provided from the CLI or a config file.
Also the purpose of the constructor is made clearer through naming 
changes.

The MockExchangeTask is flawed and not compatible to User offers,
it will soon be deprecated, so some comments were added to make that
clear.